requirements.txt
Click
Django
Flask
Flask-Cors
Flask-JWT-Extended
Jinja2
MarkupSafe
PyJWT
PyYAML
Werkzeug
asn1crypto
certifi
cffi
chardet
clickclick
connexion
cryptography
ecdsa
future
idna
inflection
itsdangerous
jose
jsonschema
numpy
openapi-spec-validator
pip
psycopg2
pyasn1
pycparser
python-jose
pytz
requests
rsa
setuptools
six
swagger-spec-validator
swagger-ui-bundle
urllib3